user_service/v2/requirements.txt
Flask==2.0.1
pymongo==3.11.0
pika==1.2.0
python-dotenv==0.19.0
